ui: add dateDividers to msg-list, use in newsfeed
