+++ ui/com/msg-list.jsx
@@ -1,5 +1,6 @@
 'use babel'
 import pull from 'pull-stream'
+import moment from 'moment'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import schemas from 'ssb-msg-schemas'
@@ -334,6 +335,7 @@ export default class MsgList extends React.Component {
     const isEmpty = (!this.state.isLoading && this.state.msgs.length === 0)
     const append = (this.state.isAtEnd && this.props.append) ? this.props.append() : ''
     const nQueued = this.state.newMsgQueue.length
+    var lastDate = moment().startOf('day')
     return <div className={'msg-list'+(this.state.selected?' msg-is-selected':'')}>
       <div className="msg-list-items">
         <Infinite
@@ -364,26 +366,19 @@ export default class MsgList extends React.Component {
             <div>
               { this.state.msgs.map((m, i) => {
                 // render item
-                return <ListItem
+                const item = <ListItem
                   key={m.key}
                   msg={m}
                   {...this.handlers}
                   selected={selectedKey === m.key}
                   forceRaw={this.props.forceRaw} />
 
-                // render "new msgs" divider, if there were new msgs
-                // TODO this had to be removed because the abstraction was wrong
-                //      the divider was meant for the newsfeed
-                //      the isNew flag tells you if the message was added since the last time the index was "touched" (read)
-                //      the problem was, msgs by the user couldnt be differentiated by the isNew item, and that was confusing
-                //      for this to work, updates by the user need to not be considered isNew
-                //      ...harder to do (than I wanted it to be) with current tooling, will need to reapproach
-                /*const wasRenderingNew = isRenderingNew
-                isRenderingNew = m.isNew
-                if (this.props.showNewDivider && wasRenderingNew && !m.isNew && i > 0) {
-                  return <div key="new-msgs-divider"><hr className="new-msgs-divider" />{item}</div>
-                }
-                return item*/
+                const oldLastDate = lastDate
+                const lastPost = threadlib.getLastThreadPost(m)
+                lastDate = moment(lastPost.value.timestamp)
+                if (this.props.dateDividers && !lastDate.isSame(oldLastDate, 'day'))
+                  return <div key={m.key}><hr className="msgs-divider" data-label={lastDate.fromNow()} />{item}</div>
+                return item
               }) }
             </div>
           }
